trace_tcp: Move tcpconnect ebpf part to its own file
